+++ skfusion/fusion/decomposition/_dfmf.py
@@ -293,7 +293,7 @@ def dfmf(R, Theta, obj_types, obj_type2rank,
 
         for r in G:
             G[r] = np.multiply(G[r], np.sqrt(
-                np.divide(G_enum[r], np.maximum(G_denom[r], np.finfo(np.float).eps))))
+                np.divide(G_enum[r], np.maximum(G_denom[r], np.finfo(float).eps))))
 
         #######################################################################
         ######################## Reconstruction Error #########################
@@ -425,7 +425,7 @@ def transform(R_ij, Theta_i, target_obj_type, obj_type2rank, G, S,
             G_enum += np.dot(theta_n, G_i)
 
         G_i = np.multiply(G_i, np.sqrt(
-            np.divide(G_enum, np.maximum(G_denom, np.finfo(np.float).eps))))
+            np.divide(G_enum, np.maximum(G_denom, np.finfo(float).eps))))
 
         #######################################################################
         ######################## Reconstruction Error #########################
